refactor: move dev into its own folder
