+++ packages/wrangler/src/dev/dev.tsx
@@ -12,21 +12,21 @@ import { withErrorBoundary, useErrorHandler } from "react-error-boundary";
 import onExit from "signal-exit";
 import tmp from "tmp-promise";
 import { fetch } from "undici";
-import { bundleWorker } from "./bundle";
-import { createWorkerPreview } from "./create-worker-preview";
-import { runCustomBuild } from "./entry";
-import useInspector from "./inspect";
-import { DEFAULT_MODULE_RULES } from "./module-collection";
-import openInBrowser from "./open-in-browser";
-import { usePreviewServer, waitForPortToBeAvailable } from "./proxy";
-import { reportError } from "./reporting";
-import { syncAssets } from "./sites";
-import { getAPIToken } from "./user";
-import type { Config } from "./config";
-import type { CfPreviewToken } from "./create-worker-preview";
-import type { Entry } from "./entry";
-import type { AssetPaths } from "./sites";
-import type { CfModule, CfWorkerInit, CfScriptFormat } from "./worker";
+import { bundleWorker } from "../bundle";
+import { createWorkerPreview } from "../create-worker-preview";
+import { runCustomBuild } from "../entry";
+import useInspector from "../inspect";
+import { DEFAULT_MODULE_RULES } from "../module-collection";
+import openInBrowser from "../open-in-browser";
+import { usePreviewServer, waitForPortToBeAvailable } from "../proxy";
+import { reportError } from "../reporting";
+import { syncAssets } from "../sites";
+import { getAPIToken } from "../user";
+import type { Config } from "../config";
+import type { CfPreviewToken } from "../create-worker-preview";
+import type { Entry } from "../entry";
+import type { AssetPaths } from "../sites";
+import type { CfModule, CfWorkerInit, CfScriptFormat } from "../worker";
 import type { WatchMode } from "esbuild";
 import type { DirectoryResult } from "tmp-promise";
